backend/requirements.txt
fastapi
google-cloud-vision
numpy
PyPDF2